projects/hitl/software/requirements.txt
ft4222
python-can==3.3.4
pyserial==3.4
pytest==6.1.1
pytest-html==2.1.1
pyusb==1.1.0